Restructure ingest status store packages
