+++ java/cdk/src/main/java/sleeper/cdk/stack/DynamoDBIngestStatusStoreStack.java
@@ -17,10 +17,10 @@
 package sleeper.cdk.stack;
 
 import sleeper.configuration.properties.InstanceProperties;
-import sleeper.ingest.job.status.DynamoDBIngestJobStatusFormat;
-import sleeper.ingest.job.status.DynamoDBIngestJobStatusStore;
-import sleeper.ingest.task.status.DynamoDBIngestTaskStatusFormat;
-import sleeper.ingest.task.status.DynamoDBIngestTaskStatusStore;
+import sleeper.ingest.status.store.job.DynamoDBIngestJobStatusFormat;
+import sleeper.ingest.status.store.job.DynamoDBIngestJobStatusStore;
+import sleeper.ingest.status.store.task.DynamoDBIngestTaskStatusFormat;
+import sleeper.ingest.status.store.task.DynamoDBIngestTaskStatusStore;
 import software.amazon.awscdk.RemovalPolicy;
 import software.amazon.awscdk.services.dynamodb.Attribute;
 import software.amazon.awscdk.services.dynamodb.AttributeType;
