+++ java/clients/src/main/java/sleeper/status/report/IngestJobStatusReport.java
@@ -26,8 +26,8 @@ import com.amazonaws.services.sqs.model.QueueAttributeName;
 import sleeper.ClientUtils;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.console.ConsoleInput;
-import sleeper.ingest.job.status.DynamoDBIngestJobStatusStore;
 import sleeper.ingest.job.status.IngestJobStatusStore;
+import sleeper.ingest.status.store.job.DynamoDBIngestJobStatusStore;
 import sleeper.job.common.CommonJobUtils;
 import sleeper.status.report.ingest.job.IngestJobStatusReportArguments;
 import sleeper.status.report.ingest.job.IngestJobStatusReporter;
