+++ java/ingest/ingest-status-store/src/test/java/sleeper/ingest/status/store/job/DynamoDBIngestJobStatusStoreCreatorIT.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.ingest.task.status;
+package sleeper.ingest.status.store.job;
 
 import com.amazonaws.services.dynamodbv2.model.DescribeTableResult;
 import org.junit.After;
@@ -24,16 +24,16 @@ import sleeper.dynamodb.tools.DynamoDBTestBase;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.ID;
-import static sleeper.ingest.testutils.IngestStatusStoreTestUtils.createInstanceProperties;
+import static sleeper.ingest.status.store.testutils.IngestStatusStoreTestUtils.createInstanceProperties;
 
-public class DynamoDBIngestTaskStatusStoreCreatorIT extends DynamoDBTestBase {
+public class DynamoDBIngestJobStatusStoreCreatorIT extends DynamoDBTestBase {
     private final InstanceProperties instanceProperties = createInstanceProperties();
-    private final String tableName = DynamoDBIngestTaskStatusStore.taskStatusTableName(instanceProperties.get(ID));
+    private final String tableName = DynamoDBIngestJobStatusStore.jobStatusTableName(instanceProperties.get(ID));
 
     @Test
     public void shouldCreateStore() {
         // When
-        DynamoDBIngestTaskStatusStoreCreator.create(instanceProperties, dynamoDBClient);
+        DynamoDBIngestJobStatusStoreCreator.create(instanceProperties, dynamoDBClient);
 
         // Then
         assertThat(dynamoDBClient.describeTable(tableName))
